Streamline manage modal more
