+++ src/components/modals/hero-state/customize-panel/customize-panel.scss
@@ -2,6 +2,7 @@
 	display: flex;
 	flex-direction: column;
 	gap: 10px;
+	padding-bottom: 20px;
 
 	.random-btn {
 		color: rgba(0, 0, 0, 0.88);
