+++ src/components/modals/hero-state/inventory-panel/inventory-panel.scss
@@ -1,0 +1,6 @@
+.inventory-panel {
+	padding: 20px 0;
+	display: flex;
+	flex-direction: column;
+	gap: 10px;	
+}
