+++ src/components/modals/hero-state/projects-panel/projects-panel.scss
@@ -1,0 +1,6 @@
+.projects-panel {
+	padding: 20px 0;
+	display: flex;
+	flex-direction: column;
+	gap: 10px;	
+}
