+++ src/components/modals/hero-state/stats-panel/stats-panel.scss
@@ -1,0 +1,6 @@
+.stats-panel {
+	padding: 20px 0;
+	display: flex;
+	flex-direction: column;
+	gap: 10px;	
+}
